+++ game/board.py
@@ -4,9 +4,15 @@ class Board():
     def __init__(self,grid=None):
         self.grid = [[ Cell(1, '') for _ in range(15) ]for _ in range(15)]
 
-    def calculate_word_value(self,word):
-        ...
-
+    def calculate_word_value(self, word):
+        value = 0
+        for cell in word:
+            value += cell.calculate_value()
+        for cell in word:
+            if cell.multiplier_type == 'word':
+                value *= cell.multiplier
+                cell.multiplier = 1
+        return value
 
 
 
@@ -26,3 +32,5 @@ class Cell:
             return 0
         if self.multiplier_type == 'letter':
             return self.letter.value * self.multiplier
+        else:
+            return self.letter.value