+++ tests/test_board.py
@@ -31,30 +31,47 @@ class TestCell(unittest.TestCase):
         self.assertEqual(cell.calculate_value(),6)
 
 
-    # def test_cell_multiplier_word(self):
-    #     word=[Cell(Tile('C',1),),Cell(Tile('A',1)),Cell(2,'word',Tile('S',2)),Cell(Tile('A',1))]
-    #     value=value.calculated_value()
-    #     self.assertEqual(value,10)
-
-    # def test_cell_multiplier_both(self):
-    #     word=[Cell(Tile('C',1),),Cell(Tile('A',1)),Cell(2,'word',Tile('S',2)),Cell(3,'letter',Tile('A',1))]
-    #     value=value.calculated_value()
-    #     self.assertEqual(value,14)
-
-    # def test_cell_no_multiplier(self):
-    #     word=[Cell(Tile('C',1),),Cell(Tile('A',1)),Cell(2,'word',Tile('S',2)),Cell(3,'letter',Tile('A',1))]
-    #     value=value.calculated_value(word)
-    #     self.assertEqual(value,14)
-    #     word=[Cell(Tile('C',1),),Cell(Tile('A',1)),Cell(2,'word',Tile('S',2)),Cell(3,'letter',Tile('A',1)),Cell(Tile('S',2))]
-    #     word.disable_multiplier()
-    #     value=value.calculated_value(word)
-    #     self.assertEqual(value,7)
+    def test_cell_multiplier_word(self):
+        cell_1 = Cell(multiplier=1,multiplier_type='letter')
+        cell_1.add_letter(Tile('C', 1)) 
+        cell_2 = Cell(multiplier=1,multiplier_type='letter')
+        cell_2.add_letter(Tile('A', 1))
+        cell_3 = Cell(multiplier=3,multiplier_type='word')
+        cell_3.add_letter(Tile('S', 2))
+        cell_4 = Cell(multiplier=1,multiplier_type='letter')
+        cell_4.add_letter(Tile('A', 1))
+        word = [cell_1, cell_2, cell_3, cell_4]
+        value=Board().calculate_word_value(word)
+        self.assertEqual(value,15)
 
 
+    def test_cell_multiplier_both(self):
+        cell_1 = Cell(multiplier=2,multiplier_type='letter')
+        cell_1.add_letter(Tile('C', 1)) 
+        cell_2 = Cell(multiplier=1,multiplier_type='letter')
+        cell_2.add_letter(Tile('A', 1))
+        cell_3 = Cell(multiplier=3,multiplier_type='word')
+        cell_3.add_letter(Tile('S', 2))
+        cell_4 = Cell(multiplier=1,multiplier_type='letter')
+        cell_4.add_letter(Tile('A', 1))
+        word = [cell_1, cell_2, cell_3, cell_4]
+        value=Board().calculate_word_value(word)
+        self.assertEqual(value,18)
 
-        
-    
-
+    def test_cell_multiplier_none(self):
+        cell_1 = Cell(multiplier=1,multiplier_type='letter')
+        cell_1.add_letter(Tile('C', 1)) 
+        cell_2 = Cell(multiplier=1,multiplier_type='letter')
+        cell_2.add_letter(Tile('A', 1))
+        cell_3 = Cell(multiplier=3,multiplier_type='word')
+        cell_3.add_letter(Tile('S', 2))
+        cell_4 = Cell(multiplier=1,multiplier_type='letter')
+        cell_4.add_letter(Tile('A', 1))
+        word = [cell_1, cell_2, cell_3, cell_4]
+        value=Board().calculate_word_value(word)
+        self.assertEqual(value,15)
+        value2=Board().calculate_word_value(word)
+        self.assertEqual(value2,5)
 
 
 if __name__ == '__main__':